requirements.txt
bs4
pyexiftool
urllib
